+++ src/common/Helpers.cs
@@ -4,7 +4,7 @@ using System.Drawing.Imaging;
 using System.IO;
 using System.Xml.Serialization;
 
-namespace common
+namespace Iris.Common
 {
     public static class Helpers
     {
