+++ src/server/Program.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
 
-namespace server
+namespace Iris.Server
 {
     static class Program
     {
